+++ Classes/Command/DeleteExpiredCommand.php
@@ -51,8 +51,9 @@ class DeleteExpiredCommand extends Command
      *
      * @param InputInterface $input
      * @param OutputInterface $output
+     * @return int
      */
-    protected function execute(InputInterface $input, OutputInterface $output)
+    protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $amountOfFiles = 0;